refactor: split fullscreen into module -> layouts, controls, lyrics/background hooks
